Finish

+++ src/main/java/lab4/lab4/dto/response/ResponseHitDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ResponseHitDTO {
-    private int x;
+    private float x;
 
     private float y;
